pyproject.toml
[tool.poetry]
authors = ["Romain Rincé"]
description = "Récupération, analyses et visualisations des rapports d'instituts de sondages."
name = "mesure_instituts_sondage"
version = "0.1.0"

[tool.poetry.dependencies]
PyPDF2 = "^1.26.0"
camelot-py = "^0.8.2"
jupyter = "^1.0.0"
matplotlib = "^3.4.1"
opencv-python-headless = "^4.5.1"
pandas = "^1.2.4"
python = ">= 3.7.1, < 3.10"
scikit-learn = "^0.24.1"
scipy = "^1.6.2"
seaborn = "^0.11.1"
statsmodels = "^0.12.2"
ing-theme-matplotlib = "^0.1.7"

[tool.poetry.dev-dependencies]

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]